refactor(matcher): update time type from int to float in barefoot matcher
Update the `time` field in `_PointMessage` and the `_last_sent_time_ms` attribute in `BarefootMatcher` from `int` to `float`.

This change allows for higher precision in timestamps by avoiding unnecessary integer truncation when converting GPS point timestamps to milliseconds. This ensures that sub-millisecond data is preserved during the matching process.

+++ mmlib/matcher/online/barefoot/_types.py
@@ -2,7 +2,7 @@ from typing import TypedDict
 
 class _PointMessage(TypedDict):
     id: str        # Identificador único do veículo/objeto
-    time: int      # Timestamp (geralmente milissegundos)
+    time: float      # Timestamp (geralmente milissegundos)
     point: str     # Formato WKT: "POINT(longitude latitude)"
 
 class _Candidate(TypedDict):
